Moved (and improved) SyncPtr to utils
